JPA & code cleanup for latest DB models

+++ jacq-common/src/main/java/org/jacq/common/model/jpa/TblImportProperties.java
@@ -50,9 +50,9 @@ public class TblImportProperties implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "IDPflanze")
-    private Integer iDPflanze;
+    private Long iDPflanze;
     @Size(max = 255)
     @Column(name = "species_name")
     private String speciesName;
@@ -67,23 +67,23 @@ public class TblImportProperties implements Serializable {
     public TblImportProperties() {
     }
 
-    public TblImportProperties(Integer id) {
+    public TblImportProperties(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getIDPflanze() {
+    public Long getIDPflanze() {
         return iDPflanze;
     }
 
-    public void setIDPflanze(Integer iDPflanze) {
+    public void setIDPflanze(Long iDPflanze) {
         this.iDPflanze = iDPflanze;
     }
